refactor: move extractProps and mergeAllOfObjectSchemas functions to build.ts
