+++ packages/serverless-typespec-generator/src/typespec/ir/convert.ts
@@ -1,49 +1,5 @@
-import { NotImplementedError } from "./error"
-import type { JSONSchema, PropIR, PropTypeIR } from "./type"
-
-export function extractProps(schema: JSONSchema): Record<string, PropIR> {
-  if (schema.allOf) {
-    return mergeAllOfObjectSchemas(schema.allOf)
-  }
-
-  const required = new Set(
-    Array.isArray(schema.required) ? schema.required : [],
-  )
-  const props: Record<string, PropIR> = {}
-
-  for (const [key, def] of Object.entries(schema.properties || {})) {
-    props[key] = {
-      type: convertType(def),
-      required: required.has(key),
-    }
-  }
-
-  return props
-}
-
-function mergeAllOfObjectSchemas(allOf: JSONSchema[]): Record<string, PropIR> {
-  const required = new Set<string>()
-  let props: Record<string, PropIR> = {}
-  for (const subSchema of allOf) {
-    if (subSchema.type !== "object") {
-      throw new NotImplementedError(
-        `Unsupported schema type in allOf: ${subSchema.type}`,
-      )
-    }
-    if (Array.isArray(subSchema.required)) {
-      for (const key of subSchema.required) {
-        required.add(key)
-      }
-    }
-    props = { ...props, ...extractProps(subSchema) }
-  }
-  for (const key of required) {
-    if (key in props) {
-      props[key].required = true
-    }
-  }
-  return props
-}
+import { extractProps } from "../../ir/build"
+import type { JSONSchema, PropTypeIR } from "./type"
 
 export function convertType(schema: JSONSchema): PropTypeIR {
   if (schema.oneOf) {
